+++ WireCalling/Sources/WireCallingUI/Views/WireMeetings/MeetingsView/MeetingsView.swift
@@ -122,121 +122,6 @@ private struct GroupedSections: View {
     }
 }
 
-// MARK: - Row
-
-private struct MeetingRow: View {
-    private typealias Strings = L10n.Localizable.WireMeetings.List
-
-    let meeting: Meeting
-    let formatTimeRange: (Meeting) -> String
-    let onEdit: () -> Void
-    let onDelete: () -> Void
-    var body: some View {
-        HStack(alignment: .top, spacing: 12) {
-            ZStack {
-                RoundedRectangle(cornerRadius: 10, style: .continuous)
-                    .fill(ColorTheme.Backgrounds.surface.color)
-                    .overlay(
-                        RoundedRectangle(cornerRadius: 10, style: .continuous)
-                            .strokeBorder(ColorTheme.Strokes.outline.color, lineWidth: 1)
-                    )
-                    .frame(width: 31, height: 31)
-
-                Image(systemName: "calendar")
-                    .font(.system(size: 15))
-            }
-
-            VStack(alignment: .leading, spacing: 2) {
-                HStack(alignment: .firstTextBaseline) {
-                    Text(meeting.title)
-                        .font(for: .body2)
-                        .foregroundStyle(ColorTheme.Backgrounds.onSurface.color)
-                        .lineLimit(2)
-
-                    Spacer()
-
-                    Menu {
-                        Button {
-                            onEdit()
-                        } label: {
-                            Label(Strings.Actions.edit, systemImage: "pencil")
-                        }
-
-                        Button(role: .destructive) {
-                            onDelete()
-                        } label: {
-                            Label(Strings.Actions.delete, systemImage: "trash")
-                        }
-                    } label: {
-                        Image(systemName: "ellipsis")
-                            .rotationEffect(.degrees(90))
-                            .foregroundStyle(ColorTheme.Buttons.Secondary.onEnabled.color)
-                            .contentShape(Rectangle())
-                    }
-                }
-
-                Text(formatTimeRange(meeting))
-                    .font(for: .subline1)
-                    .foregroundStyle(ColorTheme.Backgrounds.onSurface.color)
-
-                if !meeting.members.isEmpty {
-                    ParticipantsView(members: meeting.members)
-                        .padding(.top, 2)
-                }
-            }
-        }
-    }
-}
-
-private struct ParticipantsView: View {
-    let members: [Member]
-
-    private let maxVisible = 5
-    private let circleSize: CGFloat = 24
-    private let overlap: CGFloat = 8
-
-    var body: some View {
-        if members.count == 1 {
-            HStack(spacing: 6) {
-                circle()
-
-                if let name = members[0].name, !name.isEmpty {
-                    Text(name)
-                        .font(for: .subline1)
-                        .foregroundStyle(ColorTheme.Base.secondaryText.color)
-                        .lineLimit(1)
-                }
-            }
-        } else {
-            let overflow = members.count - maxVisible
-
-            HStack(spacing: 6) {
-                HStack(spacing: -overlap) {
-                    ForEach(Array(members.prefix(maxVisible).enumerated()), id: \.offset) { index, _ in
-                        circle()
-                            .zIndex(Double(maxVisible - index))
-                    }
-                }
-
-                if overflow > 0 {
-                    Text("+\(overflow)")
-                        .font(for: .subline1)
-                        .foregroundStyle(ColorTheme.Base.secondaryText.color)
-                }
-            }
-        }
-    }
-
-    private func circle() -> some View {
-        Circle()
-            .fill(Color.gray.opacity(0.35))
-            .frame(width: circleSize, height: circleSize)
-            .overlay(
-                Circle().strokeBorder(ColorTheme.Backgrounds.surface.color, lineWidth: 2)
-            )
-    }
-}
-
 #Preview {
     MeetingsView(viewModel: MeetingsViewModel(
         currentDateProvider: .system,
