+++ WireCalling/Tests/WireCallingTests/WireMeetings/PresentationTests/Views/MeetingsViewSnapshotTests.swift
@@ -87,7 +87,6 @@ final class MeetingsViewSnapshotTests: XCTestCase {
         )
 
         return MeetingsViewModel(
-            repository: mockRepository,
             currentDateProvider: .system,
             formatter: MeetingsFormatter(),
             upcomingMeetingsUseCase: upcomingMeetingsUseCase
